pyproject.toml
[project]
name = "music-royalty-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.65",
    "dbt-snowflake>=1.11.3",
    "python-dotenv>=1.2.2",
]
